src/constitutive/hyperelastic/test.rs
macro_rules! test_hyperelastic_constitutive_model
{
    ($hyperelastic_constitutive_model: ident, $hyperelastic_constitutive_model_parameters: expr, $hyperelastic_constitutive_model_constructed: expr) =>
    {
        use crate::constitutive::hyperelastic::HyperelasticConstitutiveModel;
        fn get_hyperelastic_constitutive_model<'a>() -> $hyperelastic_constitutive_model<'a>
        {
            $hyperelastic_constitutive_model::new($hyperelastic_constitutive_model_parameters)
        }
        #[test]
        fn get_bulk_modulus()
        {
            assert_eq!(&$hyperelastic_constitutive_model_parameters[0], get_hyperelastic_constitutive_model().get_bulk_modulus())
        }
        #[test]
        fn get_shear_modulus()
        {
            assert_eq!(get_hyperelastic_constitutive_model().get_shear_modulus(), &$hyperelastic_constitutive_model_parameters[1])
        }
        #[test]
        fn bulk_modulus()
        {
            let model = get_hyperelastic_constitutive_model();
            let deformation_gradient = DeformationGradient::identity() * (1.0 + EPSILON).powf(1.0/3.0);
            let first_piola_kirchoff_stress = model.calculate_first_piola_kirchoff_stress(&deformation_gradient);
            assert!((3.0*EPSILON*model.get_bulk_modulus()/first_piola_kirchoff_stress.trace() - 1.0).abs() < 3.0 * EPSILON);
        }
        #[test]
        fn shear_modulus()
        {
            let model = get_hyperelastic_constitutive_model();
            let mut deformation_gradient = DeformationGradient::identity();
            deformation_gradient[0][1] = EPSILON;
            let first_piola_kirchoff_stress = model.calculate_first_piola_kirchoff_stress(&deformation_gradient);
            assert!((EPSILON*model.get_shear_modulus()/first_piola_kirchoff_stress[0][1] - 1.0).abs() < EPSILON)
        }
        crate::constitutive::hyperelastic::test::test_hyperelastic_constitutive_model_constructed!($hyperelastic_constitutive_model_constructed);
    }
}
pub(crate) use test_hyperelastic_constitutive_model;
macro_rules! test_hyperelastic_constitutive_model_constructed
{
    ($hyperelastic_constitutive_model_constructed: expr) =>
    {
        use crate::
        {
            EPSILON,
            constitutive::
            {
                CauchyTangentStiffness,
                ConstitutiveModel,
                DeformationGradient,
                FirstPiolaKirchoffStress,
                FirstPiolaKirchoffTangentStiffness
            },
            math::
            {
                ContractAllIndicesWithFirstIndicesOf,
                TensorRank2Trait
            },
            mechanics::test::
            {
                get_deformation_gradient,
                get_deformation_gradient_rotated,
                get_rotation_current_configuration,
                get_rotation_reference_configuration
            },
            test::assert_eq_within_tols
        };
        fn calculate_cauchy_tangent_stiffness_from_finite_difference_of_cauchy_stress(is_deformed: bool) -> CauchyTangentStiffness
        {
            let model = $hyperelastic_constitutive_model_constructed;
            let mut cauchy_tangent_stiffness = CauchyTangentStiffness::zero();
            for k in 0..3
            {
                for l in 0..3
                {
                    let mut deformation_gradient_plus = 
                        if is_deformed
                        {
                            get_deformation_gradient()
                        }
                        else
                        {
                            DeformationGradient::identity()
                        };
                    deformation_gradient_plus[k][l] += 0.5*EPSILON;
                    let calculate_cauchy_stress_plus = model.calculate_cauchy_stress(&deformation_gradient_plus);
                    let mut deformation_gradient_minus = 
                        if is_deformed
                        {
                            get_deformation_gradient()
                        }
                        else
                        {
                            DeformationGradient::identity()
                        };
                    deformation_gradient_minus[k][l] -= 0.5*EPSILON;
                    let calculate_cauchy_stress_minus = model.calculate_cauchy_stress(&deformation_gradient_minus);
                    for i in 0..3
                    {
                        for j in 0..3
                        {
                            cauchy_tangent_stiffness[i][j][k][l] = (calculate_cauchy_stress_plus[i][j] - calculate_cauchy_stress_minus[i][j])/EPSILON;
                        }
                    }
                }
            }
            cauchy_tangent_stiffness
        }
        fn calculate_first_piola_kirchoff_stress_from_finite_difference_of_helmholtz_free_energy_density(is_deformed: bool) -> FirstPiolaKirchoffStress
        {
            let model = $hyperelastic_constitutive_model_constructed;
            let mut first_piola_kirchoff_stress = FirstPiolaKirchoffStress::zero();
            for i in 0..3
            {
                for j in 0..3
                {
                    let mut deformation_gradient_plus = 
                        if is_deformed
                        {
                            get_deformation_gradient()
                        }
                        else
                        {
                            DeformationGradient::identity()
                        };
                    deformation_gradient_plus[i][j] += 0.5*EPSILON;
                    let helmholtz_free_energy_density_plus = model.calculate_helmholtz_free_energy_density(&deformation_gradient_plus);
                    let mut deformation_gradient_minus = 
                        if is_deformed
                        {
                            get_deformation_gradient()
                        }
                        else
                        {
                            DeformationGradient::identity()
                        };
                    deformation_gradient_minus[i][j] -= 0.5*EPSILON;
                    let helmholtz_free_energy_density_minus = model.calculate_helmholtz_free_energy_density(&deformation_gradient_minus);
                    first_piola_kirchoff_stress[i][j] = (helmholtz_free_energy_density_plus - helmholtz_free_energy_density_minus)/EPSILON;
                }
            }
            first_piola_kirchoff_stress
        }
        fn calculate_first_piola_kirchoff_tangent_stiffness_from_finite_difference_of_first_piola_kirchoff_stress(is_deformed: bool) -> FirstPiolaKirchoffTangentStiffness
        {
            let model = $hyperelastic_constitutive_model_constructed;
            let mut first_piola_kirchoff_tangent_stiffness = FirstPiolaKirchoffTangentStiffness::zero();
            for k in 0..3
            {
                for l in 0..3
                {
                    let mut deformation_gradient_plus = 
                        if is_deformed
                        {
                            get_deformation_gradient()
                        }
                        else
                        {
                            DeformationGradient::identity()
                        };
                    deformation_gradient_plus[k][l] += 0.5*EPSILON;
                    let first_piola_kirchoff_stress_plus = model.calculate_first_piola_kirchoff_stress(&deformation_gradient_plus);
                    let mut deformation_gradient_minus = 
                        if is_deformed
                        {
                            get_deformation_gradient()
                        }
                        else
                        {
                            DeformationGradient::identity()
                        };
                    deformation_gradient_minus[k][l] -= 0.5*EPSILON;
                    let first_piola_kirchoff_stress_minus = model.calculate_first_piola_kirchoff_stress(&deformation_gradient_minus);
                    for i in 0..3
                    {
                        for j in 0..3
                        {
                            first_piola_kirchoff_tangent_stiffness[i][j][k][l] = (first_piola_kirchoff_stress_plus[i][j] - first_piola_kirchoff_stress_minus[i][j])/EPSILON;
                        }
                    }
                }
            }
            first_piola_kirchoff_tangent_stiffness
        }
        mod cauchy_stress
        {
            use super::*;
            mod deformed
            {
                use super::*;
                #[test]
                fn finite_difference()
                {
                    $hyperelastic_constitutive_model_constructed.calculate_cauchy_tangent_stiffness(&get_deformation_gradient()).iter()
                    .zip(calculate_cauchy_tangent_stiffness_from_finite_difference_of_cauchy_stress(true).iter())
                    .for_each(|(cauchy_tangent_stiffness_i, fd_cauchy_tangent_stiffness_i)|
                        cauchy_tangent_stiffness_i.iter()
                        .zip(fd_cauchy_tangent_stiffness_i.iter())
                        .for_each(|(cauchy_tangent_stiffness_ij, fd_cauchy_tangent_stiffness_ij)|
                            cauchy_tangent_stiffness_ij.iter()
                            .zip(fd_cauchy_tangent_stiffness_ij.iter())
                            .for_each(|(cauchy_tangent_stiffness_ijk, fd_cauchy_tangent_stiffness_ijk)|
                                cauchy_tangent_stiffness_ijk.iter()
                                .zip(fd_cauchy_tangent_stiffness_ijk.iter())
                                .for_each(|(cauchy_tangent_stiffness_ijkl, fd_cauchy_tangent_stiffness_ijkl)|
                                    assert!((cauchy_tangent_stiffness_ijkl/fd_cauchy_tangent_stiffness_ijkl - 1.0).abs() < EPSILON)
                                )
                            )
                        )
                    )
                }
                #[test]
                fn objectivity()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    model.calculate_cauchy_stress(&get_deformation_gradient()).iter()
                    .zip((
                        get_rotation_current_configuration().transpose() *
                        model.calculate_cauchy_stress(&get_deformation_gradient_rotated()) *
                        get_rotation_current_configuration()
                    ).iter())
                    .for_each(|(cauchy_stress_i, rotated_cauchy_stress_i)|
                        cauchy_stress_i.iter()
                        .zip(rotated_cauchy_stress_i.iter())
                        .for_each(|(cauchy_stress_ij, rotated_cauchy_stress_ij)|
                            assert_eq_within_tols(cauchy_stress_ij, rotated_cauchy_stress_ij)
                        )
                    )
                }
                #[test]
                fn symmetry()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    let cauchy_stress = model.calculate_cauchy_stress(&get_deformation_gradient());
                    for i in 0..3
                    {
                        for j in 0..3
                        {
                            assert_eq_within_tols(
                                &cauchy_stress[i][j],
                                &cauchy_stress[j][i]
                            )
                        }
                    }
                }
            }
            mod undeformed
            {
                use super::*;
                #[test]
                fn finite_difference()
                {
                    $hyperelastic_constitutive_model_constructed.calculate_cauchy_tangent_stiffness(&DeformationGradient::identity()).iter()
                    .zip(calculate_cauchy_tangent_stiffness_from_finite_difference_of_cauchy_stress(false).iter())
                    .for_each(|(cauchy_tangent_stiffness_i, fd_cauchy_tangent_stiffness_i)|
                        cauchy_tangent_stiffness_i.iter()
                        .zip(fd_cauchy_tangent_stiffness_i.iter())
                        .for_each(|(cauchy_tangent_stiffness_ij, fd_cauchy_tangent_stiffness_ij)|
                            cauchy_tangent_stiffness_ij.iter()
                            .zip(fd_cauchy_tangent_stiffness_ij.iter())
                            .for_each(|(cauchy_tangent_stiffness_ijk, fd_cauchy_tangent_stiffness_ijk)|
                                cauchy_tangent_stiffness_ijk.iter()
                                .zip(fd_cauchy_tangent_stiffness_ijk.iter())
                                .for_each(|(cauchy_tangent_stiffness_ijkl, fd_cauchy_tangent_stiffness_ijkl)|
                                    assert!(
                                        (cauchy_tangent_stiffness_ijkl/fd_cauchy_tangent_stiffness_ijkl - 1.0).abs() < EPSILON ||
                                        fd_cauchy_tangent_stiffness_ijkl.abs() < EPSILON
                                    )
                                )
                            )
                        )
                    )
                }
                #[test]
                fn zero()
                {
                    $hyperelastic_constitutive_model_constructed.calculate_cauchy_stress(&DeformationGradient::identity()).iter()
                    .for_each(|cauchy_stress_i|
                        cauchy_stress_i.iter()
                        .for_each(|cauchy_stress_ij|
                            assert_eq!(cauchy_stress_ij, &0.0)
                        )
                    )
                }
            }
        }
        mod cauchy_tangent_stiffness
        {
            use super::*;
            mod deformed
            {
                use super::*;
                #[test]
                fn objectivity()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    model.calculate_cauchy_tangent_stiffness(&get_deformation_gradient()).iter()
                    .zip((
                        model.calculate_cauchy_tangent_stiffness(&get_deformation_gradient_rotated())
                        .contract_all_indices_with_first_indices_of(
                            &get_rotation_current_configuration(),
                            &get_rotation_current_configuration(),
                            &get_rotation_current_configuration(),
                            &get_rotation_reference_configuration()
                        )
                    ).iter())
                    .for_each(|(cauchy_tangent_stiffness_i, rotated_cauchy_tangent_stiffness_i)|
                        cauchy_tangent_stiffness_i.iter()
                        .zip(rotated_cauchy_tangent_stiffness_i.iter())
                        .for_each(|(cauchy_tangent_stiffness_ij, rotated_cauchy_tangent_stiffness_ij)|
                            cauchy_tangent_stiffness_ij.iter()
                            .zip(rotated_cauchy_tangent_stiffness_ij.iter())
                            .for_each(|(cauchy_tangent_stiffness_ijk, rotated_cauchy_tangent_stiffness_ijk)|
                                cauchy_tangent_stiffness_ijk.iter()
                                .zip(rotated_cauchy_tangent_stiffness_ijk.iter())
                                .for_each(|(cauchy_tangent_stiffness_ijkl, rotated_cauchy_tangent_stiffness_ijkl)|
                                    assert_eq_within_tols(cauchy_tangent_stiffness_ijkl, rotated_cauchy_tangent_stiffness_ijkl)
                                )
                            )
                        )
                    )
                }
                #[test]
                fn symmetry()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    let cauchy_tangent_stiffness = model.calculate_cauchy_tangent_stiffness(&get_deformation_gradient());
                    for i in 0..3
                    {
                        for j in 0..3
                        {
                            for k in 0..3
                            {
                                for l in 0..3
                                {
                                    assert_eq_within_tols(
                                        &cauchy_tangent_stiffness[i][j][k][l],
                                        &cauchy_tangent_stiffness[j][i][k][l]
                                    )
                                }
                            }
                        }
                    }
                }
            }
            mod undeformed
            {
                use super::*;
                #[test]
                fn symmetry()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    let cauchy_tangent_stiffness = model.calculate_cauchy_tangent_stiffness(&DeformationGradient::identity());
                    for i in 0..3
                    {
                        for j in 0..3
                        {
                            for k in 0..3
                            {
                                for l in 0..3
                                {
                                    assert_eq!(cauchy_tangent_stiffness[i][j][k][l], cauchy_tangent_stiffness[j][i][k][l])
                                }
                            }
                        }
                    }
                }
            }

        }
        mod helmholtz_free_energy_density
        {
            use super::*;
            mod deformed
            {
                use super::*;
                #[test]
                fn finite_difference()
                {
                    $hyperelastic_constitutive_model_constructed.calculate_first_piola_kirchoff_stress(&get_deformation_gradient()).iter()
                    .zip(calculate_first_piola_kirchoff_stress_from_finite_difference_of_helmholtz_free_energy_density(true).iter())
                    .for_each(|(first_piola_kirchoff_stress_i, fd_first_piola_kirchoff_stress_i)|
                        first_piola_kirchoff_stress_i.iter()
                        .zip(fd_first_piola_kirchoff_stress_i.iter())
                        .for_each(|(first_piola_kirchoff_stress_ij, fd_first_piola_kirchoff_stress_ij)|
                            assert!((first_piola_kirchoff_stress_ij/fd_first_piola_kirchoff_stress_ij - 1.0).abs() < EPSILON)
                        )
                    )
                }
                #[test]
                fn objectivity()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    assert_eq_within_tols(
                        &model.calculate_helmholtz_free_energy_density(&get_deformation_gradient()),
                        &model.calculate_helmholtz_free_energy_density(&get_deformation_gradient_rotated())
                    )
                }
                #[test]
                fn positive()
                {
                    assert!($hyperelastic_constitutive_model_constructed.calculate_helmholtz_free_energy_density(&get_deformation_gradient()) > 0.0)
                }
            }
            mod undeformed
            {
                use super::*;
                #[test]
                fn finite_difference()
                {
                    calculate_first_piola_kirchoff_stress_from_finite_difference_of_helmholtz_free_energy_density(false).iter()
                    .for_each(|fd_first_piola_kirchoff_stress_i|
                        fd_first_piola_kirchoff_stress_i.iter()
                        .for_each(|fd_first_piola_kirchoff_stress_ij|
                            assert!(fd_first_piola_kirchoff_stress_ij.abs() < EPSILON)
                        )
                    )
                }
                #[test]
                fn zero()
                {
                    assert_eq!($hyperelastic_constitutive_model_constructed.calculate_helmholtz_free_energy_density(&DeformationGradient::identity()), 0.0)
                }
            }
        }
        mod first_piola_kirchoff_stress
        {
            use super::*;
            mod deformed
            {
                use super::*;
                #[test]
                fn finite_difference()
                {
                    $hyperelastic_constitutive_model_constructed.calculate_first_piola_kirchoff_tangent_stiffness(&get_deformation_gradient()).iter()
                    .zip(calculate_first_piola_kirchoff_tangent_stiffness_from_finite_difference_of_first_piola_kirchoff_stress(true).iter())
                    .for_each(|(first_piola_kirchoff_tangent_stiffness_i, fd_first_piola_kirchoff_tangent_stiffness_i)|
                        first_piola_kirchoff_tangent_stiffness_i.iter()
                        .zip(fd_first_piola_kirchoff_tangent_stiffness_i.iter())
                        .for_each(|(first_piola_kirchoff_tangent_stiffness_ij, fd_first_piola_kirchoff_tangent_stiffness_ij)|
                            first_piola_kirchoff_tangent_stiffness_ij.iter()
                            .zip(fd_first_piola_kirchoff_tangent_stiffness_ij.iter())
                            .for_each(|(first_piola_kirchoff_tangent_stiffness_ijk, fd_first_piola_kirchoff_tangent_stiffness_ijk)|
                                first_piola_kirchoff_tangent_stiffness_ijk.iter()
                                .zip(fd_first_piola_kirchoff_tangent_stiffness_ijk.iter())
                                .for_each(|(first_piola_kirchoff_tangent_stiffness_ijkl, fd_first_piola_kirchoff_tangent_stiffness_ijkl)|
                                    assert!((first_piola_kirchoff_tangent_stiffness_ijkl/fd_first_piola_kirchoff_tangent_stiffness_ijkl - 1.0).abs() < EPSILON)
                                )
                            )
                        )
                    )
                }
                #[test]
                fn objectivity()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    model.calculate_first_piola_kirchoff_stress(&get_deformation_gradient()).iter()
                    .zip((
                        get_rotation_current_configuration().transpose() *
                        model.calculate_first_piola_kirchoff_stress(&get_deformation_gradient_rotated()) *
                        get_rotation_reference_configuration()
                    ).iter())
                    .for_each(|(first_piola_kirchoff_stress_i, rotated_first_piola_kirchoff_stress_i)|
                        first_piola_kirchoff_stress_i.iter()
                        .zip(rotated_first_piola_kirchoff_stress_i.iter())
                        .for_each(|(first_piola_kirchoff_stress_ij, rotated_first_piola_kirchoff_stress_ij)|
                            assert_eq_within_tols(first_piola_kirchoff_stress_ij, rotated_first_piola_kirchoff_stress_ij)
                        )
                    )
                }
            }
            mod undeformed
            {
                use super::*;
                #[test]
                fn finite_difference()
                {
                    $hyperelastic_constitutive_model_constructed.calculate_first_piola_kirchoff_tangent_stiffness(&DeformationGradient::identity()).iter()
                    .zip(calculate_first_piola_kirchoff_tangent_stiffness_from_finite_difference_of_first_piola_kirchoff_stress(false).iter())
                    .for_each(|(first_piola_kirchoff_tangent_stiffness_i, fd_first_piola_kirchoff_tangent_stiffness_i)|
                        first_piola_kirchoff_tangent_stiffness_i.iter()
                        .zip(fd_first_piola_kirchoff_tangent_stiffness_i.iter())
                        .for_each(|(first_piola_kirchoff_tangent_stiffness_ij, fd_first_piola_kirchoff_tangent_stiffness_ij)|
                            first_piola_kirchoff_tangent_stiffness_ij.iter()
                            .zip(fd_first_piola_kirchoff_tangent_stiffness_ij.iter())
                            .for_each(|(first_piola_kirchoff_tangent_stiffness_ijk, fd_first_piola_kirchoff_tangent_stiffness_ijk)|
                                first_piola_kirchoff_tangent_stiffness_ijk.iter()
                                .zip(fd_first_piola_kirchoff_tangent_stiffness_ijk.iter())
                                .for_each(|(first_piola_kirchoff_tangent_stiffness_ijkl, fd_first_piola_kirchoff_tangent_stiffness_ijkl)|
                                    assert!(
                                        (first_piola_kirchoff_tangent_stiffness_ijkl/fd_first_piola_kirchoff_tangent_stiffness_ijkl - 1.0).abs() < EPSILON ||
                                        fd_first_piola_kirchoff_tangent_stiffness_ijkl.abs() < EPSILON
                                    )
                                )
                            )
                        )
                    )
                }
                #[test]
                fn zero()
                {
                    $hyperelastic_constitutive_model_constructed.calculate_first_piola_kirchoff_stress(&DeformationGradient::identity()).iter()
                    .for_each(|first_piola_kirchoff_stress_i|
                        first_piola_kirchoff_stress_i.iter()
                        .for_each(|first_piola_kirchoff_stress_ij|
                            assert_eq!(first_piola_kirchoff_stress_ij, &0.0)
                        )
                    )
                }
            }
        }
        mod first_piola_kirchoff_tangent_stiffness
        {
            use super::*;
            mod deformed
            {
                use super::*;
                #[test]
                fn objectivity()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    model.calculate_first_piola_kirchoff_tangent_stiffness(&get_deformation_gradient()).iter()
                    .zip((
                        model.calculate_first_piola_kirchoff_tangent_stiffness(&get_deformation_gradient_rotated())
                        .contract_all_indices_with_first_indices_of(
                            &get_rotation_current_configuration(),
                            &get_rotation_reference_configuration(),
                            &get_rotation_current_configuration(),
                            &get_rotation_reference_configuration()
                        )
                    ).iter())
                    .for_each(|(first_piola_kirchoff_tangent_stiffness_i, rotated_first_piola_kirchoff_tangent_stiffness_i)|
                        first_piola_kirchoff_tangent_stiffness_i.iter()
                        .zip(rotated_first_piola_kirchoff_tangent_stiffness_i.iter())
                        .for_each(|(first_piola_kirchoff_tangent_stiffness_ij, rotated_first_piola_kirchoff_tangent_stiffness_ij)|
                            first_piola_kirchoff_tangent_stiffness_ij.iter()
                            .zip(rotated_first_piola_kirchoff_tangent_stiffness_ij.iter())
                            .for_each(|(first_piola_kirchoff_tangent_stiffness_ijk, rotated_first_piola_kirchoff_tangent_stiffness_ijk)|
                                first_piola_kirchoff_tangent_stiffness_ijk.iter()
                                .zip(rotated_first_piola_kirchoff_tangent_stiffness_ijk.iter())
                                .for_each(|(first_piola_kirchoff_tangent_stiffness_ijkl, rotated_first_piola_kirchoff_tangent_stiffness_ijkl)|
                                    assert_eq_within_tols(first_piola_kirchoff_tangent_stiffness_ijkl, rotated_first_piola_kirchoff_tangent_stiffness_ijkl)
                                )
                            )
                        )
                    )
                }
                #[test]
                fn symmetry()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    let first_piola_kirchoff_tangent_stiffness = model.calculate_first_piola_kirchoff_tangent_stiffness(&get_deformation_gradient());
                    for i in 0..3
                    {
                        for j in 0..3
                        {
                            for k in 0..3
                            {
                                for l in 0..3
                                {
                                    assert_eq_within_tols(
                                        &first_piola_kirchoff_tangent_stiffness[i][j][k][l],
                                        &first_piola_kirchoff_tangent_stiffness[k][l][i][j]
                                    )
                                }
                            }
                        }
                    }
                }
            }
            mod undeformed
            {
                use super::*;
                #[test]
                fn symmetry()
                {
                    let model = $hyperelastic_constitutive_model_constructed;
                    let first_piola_kirchoff_tangent_stiffness = model.calculate_first_piola_kirchoff_tangent_stiffness(&DeformationGradient::identity());
                    for i in 0..3
                    {
                        for j in 0..3
                        {
                            for k in 0..3
                            {
                                for l in 0..3
                                {
                                    assert_eq_within_tols(
                                        &first_piola_kirchoff_tangent_stiffness[i][j][k][l],
                                        &first_piola_kirchoff_tangent_stiffness[k][l][i][j]
                                    )
                                }
                            }
                        }
                    }
                }
            }

        }
    }
}
pub(crate) use test_hyperelastic_constitutive_model_constructed;